+++ app/services/places/name_fetcher.rb
@@ -12,7 +12,8 @@ module Places
       ).first
       return nil if result.blank?
 
-      properties = result.data&.dig('properties')
+      properties = result.data.is_a?(Hash) ? result.data['properties'] : nil
+      properties ||= Geocoding::ResultNormalizer.call(result)[:properties]
       return nil if properties.blank?
 
       name = ::Visits::Names::Builder.build_from_properties(properties)